Introduce Github actions workflow file
Resolves dependencies by using a couple of scripts from qt5 repo:
one to list the yaml file in a "$repo $ref" format, and another to
download each needed artifact using the Github REST API.
Also uploads build artifacts for other modules to fetch.

Currently it is set to fetch the latest build artifacts for each
dependency, instead of the specific ref listed in dependencies.yaml. The
reason is two-fold:

1. it is necessary for now because older build artifacts were never
   generated
2. it has value by itself, doing a quick check on whether the module is
   broken against qtbase.

In the future we could easily add a 2nd build configuration by removing
the FORCE_BRANCH variable.

Finally, this config has nothing specific to this module, it can be used
in other submodules too, as-is or with minor changes. It is being
duplicated across modules however, until fully composite github actions
are supported:
https://github.com/actions/runner/issues/646

Change-Id: I491cf189e9b777eaf7edcee9ae55355d583ac03b